Move remaining test files to providers
