+++ client/src/components/providers/History/caching/db/find.test.js
@@ -8,7 +8,7 @@ import { content$, dscContent$, buildContentId } from "./observables";
 import { find } from "./find";
 
 // test data
-import historyContent from "components/History/test/json/historyContent.json";
+import historyContent from "components/providers/History/test/json/historyContent.json";
 import collectionContent from "components/providers/History/test/json/collectionContent.json";
 
 beforeEach(wipeDatabase);
