+++ client/src/components/providers/History/caching/db/monitorQuery.test.js
@@ -9,7 +9,7 @@ import { content$, dscContent$ } from "./observables";
 import { cacheContent, cacheCollectionContent, bulkCacheContent, bulkCacheDscContent } from "./promises";
 
 // test data
-import historyContent from "components/History/test/json/historyContent.json";
+import historyContent from "components/providers/History/test/json/historyContent.json";
 import collectionContent from "components/providers/History/test/json/collectionContent.json";
 
 jest.mock("app");
